requirements.txt
torch>=2.0.0
opencv-python-headless>=4.8.0
flask>=3.0.0
pillow>=9.0.0
